complete

+++ src/main/java/com/xbeast/xbeast/repository/ClienteRepository.java
@@ -6,6 +6,6 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ClienteRepository extends JpaRepository<Cliente, Long> {
+public interface ClienteRepository extends JpaRepository<Cliente, Integer> {
     UserDetails findByEmail(String email);
-} 
+}
